roams_backend/requirements.txt
aiofiles==24.1.0
aiosqlite==0.21.0
amqp==5.3.1
asgiref==3.9.1
asyncua==1.1.6
billiard==4.2.1
celery==5.5.3
cffi==1.17.1
channels==4.3.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
concurrent-log-handler==0.9.28
cryptography==45.0.7
Django==4.2.23
django-cors-headers==4.7.0
django-filter==24.1
django-redis==6.0.0
djangorestframework==3.16.1
h5py==3.14.0
kombu==5.5.4
lockfile==0.12.2
lxml==6.0.1
MonthDelta==0.9.1
numpy==2.3.2
opcua==0.98.13
opcua-client==0.8.4
opcua-widgets==0.6.1
packaging==25.0
paho-mqtt==2.1.0
pillow==11.3.0
portalocker==3.2.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.22
pyOpenSSL==25.1.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PyScada==0.8.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-environ==0.4.54
pytz==2025.2
redis==6.4.0
requests==2.31.0
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
